Move preview data code to previewdata file
